update db 1.9

+++ BMI_Tracker/src/main/java/com/fu/bmi_tracker/util/BMIUtils.java
@@ -26,7 +26,7 @@ public class BMIUtils {
     }
 
     // Method to calculate BMR Mifflin St Jeor
-    public double calculateBMR(double weight, double height, int age, EGender gender) {
+    public double calculateBMR(int weight, int height, int age, EGender gender) {
         double bMR;
         if (EGender.Male.equals(gender)) {
             //( 10 x weight (kg) + 6.25 x height (cm) – 5 x age (y) + 5
